27/9

+++ src/Vehicle/BikeCycle.java
@@ -1,9 +1,11 @@
 package Vehicle;
 
+import java.math.BigInteger;
+
 public class BikeCycle extends Vehicle{
 	protected String tyre;
 
-	public BikeCycle(Integer id, String brand, String type, Integer speed, Integer price, String tyre) {
+	public BikeCycle(Integer id, String brand, String type, Integer speed, BigInteger price, String tyre) {
 		super(id, brand, type, speed, price);
 		this.tyre = tyre;
 	}
